LayoutTests/svg/dynamic-updates/SVGFEOffsetElement-svgdom-dx-prop-expected.txt

CONSOLE MESSAGE: line 38: TypeError: 'undefined' is not an object (evaluating 'offsetElement.dx.baseVal = 0')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dx' property of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL offsetElement.dx.baseVal should be 50. Threw exception TypeError: 'undefined' is not an object (evaluating 'offsetElement.dx.baseVal')
